Rename pathfind/driving -> pathfind/vehicles to reflect bikes using it too, and likewise for driving_cost
